ctree: add the SubItem class
This is a helper class to identify nested structs in ItemData objects.
An example is the Chunk object, which contains Stripe objects.

I'm going to add a streaming object printer based on the __str__
methods of objects, which can quickly display objects in a more concise
way than the very verbose pretty printer. By checking if an attribute of
an ItemData object is of type SubItem, we can decide to also print it.