WIP btrfs-search-metadata: Hello, world!
After moving a bunch of interesting example programs into the bin/
location I'm looking at the rest of the random pile of code in examples/

What about a tool that allows searching any metadata you want? And what
if it already has a bunch of presets that select ranges from trees
automatically?

    btrfs-search-metadata [-l|--long] <preset> [<moar> ..] <path>

The first version of this program has preset search options for:
- chunks
- block groups
- all information for an inode (objectid) in any subvolume
- devices (the metadata items)
- dev_info (the devices info from the ioctl)
- dev_stats (also from the ioctl)
- dump: custom search query for a range of items in a tree

TODO:
- min_key and max_key for dump
- orphans, what to display?